Add test case showing failure for modules with paragraphs
