+++ importers/src/bible/osis/tests/OsisImporter.test.ts
@@ -1,34 +1,24 @@
 import { resolve } from 'path';
-import { getBibleEngineTestInstance, TEST_BIBLE_VERSION } from './utils';
-import { OsisImporter } from '../../..';
-import { enBookMetadata } from '../../../metadata/index';
+import { getContextFromSource } from './utils';
 import { IBibleContentSection } from '@bible-engine/core';
 
-const bibleEngine = getBibleEngineTestInstance();
-
 describe('OSIS Parser', () => {
-    describe('getContextFromXml', () => {
-        it('wraps each section in a paragraph, if no paragraphs exist in the source text', async () => {
-            const sourcePath = resolve(__dirname) + '/data/sectionsButNoParagraphs.xml';
-            const importer = new OsisImporter(bibleEngine, {
-                sourcePath,
-                bookMeta: enBookMetadata,
-                versionMeta: TEST_BIBLE_VERSION,
-            });
-            // try {
-            const xml = await importer.getXmlFromOptions({
-                sourcePath,
-                bookMeta: enBookMetadata,
-            });
-            const context = await importer.getContextFromXml(xml);
-            expect(context.books.length);
-            const SECTION = context.books[0].contents[0] as IBibleContentSection;
-            expect(SECTION.type).toBe('section');
-            const PARAGRAPH: any = SECTION.contents[0];
-            expect(PARAGRAPH.groupType).toBe('paragraph');
-            // } catch (error) {
-            //     console.log('caught in jest!');
-            // }
-        });
+    it('wraps each section in a paragraph, if no paragraphs exist in the source text', async () => {
+        const sourcePath = resolve(__dirname) + '/data/sectionsButNoParagraphs.xml';
+        const context = await getContextFromSource(sourcePath);
+        expect(context.books.length);
+        const SECTION = context.books[0].contents[0] as IBibleContentSection;
+        expect(SECTION.type).toBe('section');
+        const PARAGRAPH: any = SECTION.contents[0];
+        expect(PARAGRAPH.groupType).toBe('paragraph');
+    });
+    it('handles natural paragraphs and sections, when included together', async () => {
+        const sourcePath = resolve(__dirname) + '/data/sectionsPlusParagraphs.xml';
+        const context = await getContextFromSource(sourcePath);
+        expect(context.books.length);
+        const SECTION = context.books[0].contents[0] as IBibleContentSection;
+        expect(SECTION.type).toBe('section');
+        const PARAGRAPH: any = SECTION.contents[0];
+        expect(PARAGRAPH.groupType).toBe('paragraph');
     });
 });
