+++ importers/src/bible/osis/tests/utils.ts
@@ -1,5 +1,7 @@
 import { BibleEngine, IBibleContentSection } from '@bible-engine/core';
 import { ConnectionOptions } from 'typeorm';
+import { OsisImporter } from '..';
+import { enBookMetadata } from '../../../metadata';
 
 const CONNECTION_OPTIONS: ConnectionOptions = {
     type: 'sqlite',
@@ -23,4 +25,18 @@ export const getEmptySection = (): IBibleContentSection => {
         level: 0,
         contents: [],
     };
+};
+
+export const getContextFromSource = async (sourcePath: string) => {
+    const bibleEngine = getBibleEngineTestInstance();
+    const importer = new OsisImporter(bibleEngine, {
+        sourcePath,
+        bookMeta: enBookMetadata,
+        versionMeta: TEST_BIBLE_VERSION,
+    });
+    const xml = await importer.getXmlFromOptions({
+        sourcePath,
+        bookMeta: enBookMetadata,
+    });
+    return importer.getContextFromXml(xml);
 };
